requirements.txt
# Dev environment
pip
autopep8

# App
asgiref==3.7.2
Django==5.0.13
djangorestframework
sqlparse==0.4.4
typing_extensions==4.9.0
psycopg2-binary==2.9.9
python-decouple==3.8
